+++ README.md
@@ -2,7 +2,7 @@
 
 Code Coverage
 
-![Statements](https://img.shields.io/badge/statements-74.05%25-red.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-52.63%25-red.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-70.66%25-red.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-76.06%25-red.svg?style=flat)
+![Statements](https://img.shields.io/badge/statements-72.55%25-red.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-47.16%25-red.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-65.85%25-red.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-74.11%25-red.svg?style=flat)
