+++ src/interfaces/permissions/index.ts
@@ -1,6 +1,7 @@
-import { PermissionsRequest, processPermissionsRequest } from './request';
+import { PermissionsRequest } from './messages/permissions-request';
+import { handlePermissionsRequest } from './handlers/permissions-request';
 
 export const PermissionsInterface = {
-  methodHandlers : { 'PermissionsRequest': processPermissionsRequest },
+  methodHandlers : { 'PermissionsRequest': handlePermissionsRequest },
   messages       : [ PermissionsRequest ]
 };
