--- src/interfaces/permissions/request/index.ts
@@ -1,2 +1,0 @@
-export { PermissionsRequest, DEFAULT_CONDITIONS } from './message';
-export { processPermissionsRequest } from './method';
