--- src/interfaces/permissions/request/types.ts
@@ -1,16 +1,0 @@
-import type { Authorization, BaseMessageSchema } from '../../../messages/types';
-import type { Conditions, Scope } from '../types';
-
-export type PermissionsRequestDescriptor = {
-  description: string
-  grantedTo: string
-  grantedBy: string
-  method: 'PermissionsRequest'
-  scope: Scope
-  conditions: Conditions
-  objectId?: string
-};
-
-export type PermissionsRequestSchema = BaseMessageSchema & Authorization & {
-  descriptor: PermissionsRequestDescriptor;
-};
