--- src/response/index.ts
@@ -1,2 +1,0 @@
-export * from './response';
-export * from './message-reply';
